Bug 1917632 - Apply the sidebar revamp in the same window that pref is enabled in  a=dmeehan `adoptFromWindow` sets the sidebar width based on the "opener" window. If the opener does not have a sidebar, no width is remembered. Previously, flipping the sidebar revamp pref required opening a new window to see the revamped sidebar, and hence meant the opener did not have a sidebar.
This patch ensures the revamp is applied in the current/opener window and hence the width
of the sidebar opened in that window is used to determine the width of the sidebar in future
windows opened. This should resolve the issue some users have found when closing the browser
via the X button and reopening the browser doesn't remember the sidebar width.